requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.9
django-allauth==0.46.0
django-environ==0.8.1
django-import-export==2.6.1
djangorestframework==3.12.4
et-xmlfile==1.1.0
idna==3.3
MarkupPy==1.14
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.9
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.2
tablib==3.1.0
typing_extensions==4.0.0
urllib3==1.26.7
xlrd==2.0.1
xlwt==1.3.0
